Refactor: cts: Move cts.environment to _cts module.
