+++ cts/lab/ClusterManager.py
@@ -17,9 +17,9 @@ from cts.CIB         import ConfigFactory
 from cts.CTS         import NodeStatus, Process
 from cts.CTStests    import AuditResource
 from cts.watcher     import LogWatcher
-from cts.environment import EnvFactory
 
 from pacemaker.buildoptions import BuildOptions
+from pacemaker._cts.environment import EnvFactory
 from pacemaker._cts.logging import LogFactory
 from pacemaker._cts.patterns import PatternSelector
 from pacemaker._cts.remote import RemoteFactory
