+++ python/pacemaker/_cts/Makefile.am
@@ -13,6 +13,7 @@ pkgpythondir = $(pythondir)/$(PACKAGE)/_cts
 
 pkgpython_PYTHON = __init__.py \
 				   corosync.py \
+				   environment.py \
 				   errors.py \
 				   logging.py \
 				   patterns.py \
